~ changed loading animation after login/signup + fixed project details fragment able to show edit team dialog ~ LoginLandingFragment,  SignInFragment, SignUpFragment are now using TacPpFragment ~ changed app bar background color to a darker one
