+++ app/src/main/res/color/color_selector_efab_text.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/colorPrimaryDark" android:state_enabled="true"/>
+    <item android:color="@color/indigo_100_blacked_transparent" android:state_enabled="false"/>
+</selector>
