+++ app/src/main/res/drawable/transient_semitransparent_rectangle_app_bar.xml
@@ -1,7 +1,7 @@
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
     <item android:id="@+id/opaque_background">
         <shape android:shape="rectangle">
-            <solid android:color="@color/indigo.300"/>
+            <solid android:color="@color/indigo.A100_grayed"/>
             <corners
                     android:bottomRightRadius="@dimen/_list_item_corner_radius_large"
             />
@@ -15,7 +15,7 @@
                     android:bottomRightRadius="@dimen/_list_item_corner_radius_large"
             />
             <gradient
-                    android:startColor="@color/semitransparent_black_left"
+                    android:startColor="@color/semitransparent_black_left_selected"
                     android:endColor="@color/semitransparent_black_right"
             />
         </shape>
