+++ app/src/main/res/layout/fragment_tac_pp.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+    <androidx.appcompat.widget.AppCompatTextView
+            android:id="@+id/tac_text"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginLeft="16dp"
+            android:layout_marginStart="16dp"
+            android:layout_marginBottom="16dp"
+            android:textColor="@color/indigo.100"
+            android:text="@string/terms_and_conditions"
+            android:layout_gravity="bottom|start"
+    />
+
+    <androidx.appcompat.widget.AppCompatTextView
+            android:id="@+id/pp_text"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginRight="16dp"
+            android:layout_marginEnd="16dp"
+            android:layout_marginBottom="16dp"
+            android:textColor="@color/indigo.100"
+            android:text="@string/privacy_policy"
+            android:layout_gravity="bottom|end"
+    />
+</FrameLayout>
