+++ app/src/main/res/layout/layout_landing.xml
@@ -2,6 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
+        xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:id="@+id/landing_container"
@@ -68,7 +69,6 @@
             android:layout_marginLeft="@dimen/landing_button_margin"
             android:backgroundTint="@color/pureWhite"
             android:background="@color/pureWhite"
-            android:foreground="?selectableItemBackground"
             android:text="@string/sign_up"
             android:gravity="center"
             app:layout_constraintBottom_toBottomOf="parent"
@@ -84,37 +84,22 @@
             android:layout_marginRight="@dimen/landing_button_margin"
             android:backgroundTint="@color/pureWhite"
             android:background="@color/pureWhite"
-            android:foreground="?selectableItemBackground"
             android:text="@string/sign_in"
             android:gravity="center"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
     />
 
-    <androidx.appcompat.widget.AppCompatTextView
-            android:id="@+id/tac_text"
-            android:layout_width="wrap_content"
+    <fragment
+            android:id="@+id/tac_pp"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginLeft="16dp"
-            android:layout_marginStart="16dp"
-            android:layout_marginBottom="16dp"
-            android:textColor="@color/indigo.100"
-            android:text="@string/terms_and_conditions"
+            android:name="org.rares.miner49er.ui.fragments.login.TacPpFragment"
+            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-    />
-
-    <androidx.appcompat.widget.AppCompatTextView
-            android:id="@+id/pp_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginRight="16dp"
-            android:layout_marginEnd="16dp"
-            android:layout_marginBottom="16dp"
-            android:textColor="@color/indigo.100"
-            android:text="@string/privacy_policy"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
+            tools:context="org.rares.miner49er.ui.fragments.login.TacPpFragment"
+            tools:layout="@layout/fragment_tac_pp"
     />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
