+++ app/src/main/res/layout/layout_sign_in.xml
@@ -59,6 +59,7 @@
             android:gravity="center"
             android:backgroundTint="@color/pureWhite"
             android:background="@color/pureWhite"
+            android:textColor="@color/color_selector_efab_text"
             android:text="@string/login_text"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
@@ -76,20 +77,21 @@
             app:layout_constraintStart_toStartOf="@id/login_floating_button"
             app:layout_constraintEnd_toEndOf="@id/login_floating_button"
             app:layout_constraintTop_toBottomOf="@id/login_floating_button"
-            app:layout_constraintBottom_toTopOf="@+id/tac_text"
+            app:layout_constraintBottom_toTopOf="@+id/tac_pp"
             tools:text="@string/login_failed_bad_u_p"
     />
 
     <androidx.core.widget.ContentLoadingProgressBar
             android:id="@+id/progress_circular"
-            android:layout_width="64dp"
-            android:layout_height="64dp"
+            android:layout_width="0dp"
+            android:layout_height="48dp"
             android:indeterminate="true"
-            android:indeterminateTint="@color/pureWhite"
+            android:indeterminateTint="?colorOnPrimary"
             android:indeterminateDuration="100"
             android:indeterminateBehavior="repeat"
             android:background="@android:color/transparent"
             android:visibility="gone"
+            android:elevation="10dp"
             app:layout_constraintStart_toStartOf="@id/login_floating_button"
             app:layout_constraintEnd_toEndOf="@id/login_floating_button"
             app:layout_constraintTop_toTopOf="@id/login_floating_button"
@@ -97,30 +99,16 @@
             style="?android:attr/progressBarStyleLarge"
     />
 
-    <androidx.appcompat.widget.AppCompatTextView
-            android:id="@+id/tac_text"
-            android:layout_width="wrap_content"
+    <fragment
+            android:id="@+id/tac_pp"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginLeft="16dp"
-            android:layout_marginStart="16dp"
-            android:layout_marginBottom="16dp"
-            android:textColor="@color/indigo.100"
-            android:text="@string/terms_and_conditions"
+            android:name="org.rares.miner49er.ui.fragments.login.TacPpFragment"
+            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-    />
-
-    <androidx.appcompat.widget.AppCompatTextView
-            android:id="@+id/pp_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginRight="16dp"
-            android:layout_marginEnd="16dp"
-            android:layout_marginBottom="16dp"
-            android:textColor="@color/indigo.100"
-            android:text="@string/privacy_policy"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
+            tools:context="org.rares.miner49er.ui.fragments.login.TacPpFragment"
+            tools:layout="@layout/fragment_tac_pp"
     />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
