+++ app/src/main/res/layout/layout_sign_up.xml
@@ -115,16 +115,16 @@
             app:layout_constraintStart_toStartOf="@id/create_account_btn"
             app:layout_constraintEnd_toEndOf="@id/create_account_btn"
             app:layout_constraintTop_toBottomOf="@id/create_account_btn"
-            app:layout_constraintBottom_toTopOf="@+id/tac_text"
+            app:layout_constraintBottom_toTopOf="@+id/tac_pp"
             tools:text="@string/create_user_failed_user_exists"
     />
 
     <androidx.core.widget.ContentLoadingProgressBar
             android:id="@+id/progress_circular"
-            android:layout_width="64dp"
-            android:layout_height="64dp"
+            android:layout_width="0dp"
+            android:layout_height="48dp"
             android:indeterminate="true"
-            android:indeterminateTint="@color/pureWhite"
+            android:indeterminateTint="?colorOnPrimary"
             android:indeterminateDuration="100"
             android:indeterminateBehavior="repeat"
             android:background="@android:color/transparent"
@@ -137,30 +137,16 @@
             style="?android:attr/progressBarStyleLarge"
     />
 
-    <androidx.appcompat.widget.AppCompatTextView
-            android:id="@+id/tac_text"
-            android:layout_width="wrap_content"
+    <fragment
+            android:id="@+id/tac_pp"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginLeft="16dp"
-            android:layout_marginStart="16dp"
-            android:layout_marginBottom="16dp"
-            android:textColor="@color/indigo.100"
-            android:text="@string/terms_and_conditions"
+            android:name="org.rares.miner49er.ui.fragments.login.TacPpFragment"
+            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-    />
-
-    <androidx.appcompat.widget.AppCompatTextView
-            android:id="@+id/pp_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginRight="16dp"
-            android:layout_marginEnd="16dp"
-            android:layout_marginBottom="16dp"
-            android:textColor="@color/indigo.100"
-            android:text="@string/privacy_policy"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
+            tools:context="org.rares.miner49er.ui.fragments.login.TacPpFragment"
+            tools:layout="@layout/fragment_tac_pp"
     />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
